# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Awesome API Reference",
        "version": "1.0.0"
    },
    "paths": {
        "/redis/set": {
            "post": {
                "tags": [
                    "Redis"
                ],
                "summary": "Set Data",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string"
                                    },
                                    "gender": {
                                        "type": "string"
                                    },
                                    "age": {
                                        "type": "number"
                                    }
                                },
                                "type": "object",
                                "example": {
                                    "name": "Lustria Ebiz",
                                    "gender": "Male",
                                    "age": 24
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "An example resource"
                    }
                }
            }
        },
        "/redis/get": {
            "get": {
                "tags": [
                    "Redis"
                ],
                "summary": "Get Data",
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
        "/redis/publish": {
            "post": {
                "tags": [
                    "Redis"
                ],
                "summary": "Publish Data",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "channel": {
                                        "type": "string"
                                    },
                                    "message": {
                                        "type": "string"
                                    }
                                },
                                "type": "object",
                                "example": {
                                    "channel": "chanel-1022",
                                    "message": "Hay bro"
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "An example resource"
                    }
                }
            }
        },
        "/redis/destroy": {
            "delete": {
                "tags": [
                    "Redis"
                ],
                "summary": "Delete Data",
                "responses": {
                    "200": {
                        "description": "An example resource"
                    }
                }
            }
        },
        "/roles/list": {
            "get": {
                "tags": [
                    "Roles"
                ],
                "summary": "Get all roles",
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                },
                "security": [
                    {
                        "apiAuth": []
                    }
                ]
            }
        },
        "/roles/create-with-permissions": {
            "post": {
                "tags": [
                    "Roles"
                ],
                "summary": "Create Roles & Permission",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "role": {
                                        "type": "string"
                                    },
                                    "permissions": {
                                        "type": "string"
                                    }
                                },
                                "type": "object",
                                "example": {
                                    "role": "subscriber",
                                    "permissions": [
                                        "view_article",
                                        "like_article"
                                    ]
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "An example resource"
                    }
                },
                "security": [
                    {
                        "apiAuth": []
                    }
                ]
            }
        },
        "/user/role-permissions": {
            "get": {
                "tags": [
                    "User"
                ],
                "summary": "User Has Role Permissions",
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                },
                "security": [
                    {
                        "apiAuth": []
                    }
                ]
            }
        },
        "/user/assign-roles": {
            "post": {
                "tags": [
                    "User"
                ],
                "summary": "Assign roles to user",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "roles": {
                                        "type": "string"
                                    }
                                },
                                "type": "object",
                                "example": {
                                    "roles": [
                                        "editor"
                                    ]
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "An example resource"
                    }
                },
                "security": [
                    {
                        "apiAuth": []
                    }
                ]
            }
        },
        "/user/remove-role": {
            "post": {
                "tags": [
                    "User"
                ],
                "summary": "Rovoke permissions to user",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "roles": {
                                        "type": "string"
                                    }
                                },
                                "type": "object",
                                "example": {
                                    "roles": "editor"
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "An example resource"
                    }
                },
                "security": [
                    {
                        "apiAuth": []
                    }
                ]
            }
        },
        "/user/give-permissions": {
            "post": {
                "tags": [
                    "User"
                ],
                "summary": "Give permissions to user",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "permission": {
                                        "type": "string"
                                    }
                                },
                                "type": "object",
                                "example": {
                                    "permission": [
                                        "edit_article",
                                        "view__article"
                                    ]
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "An example resource"
                    }
                },
                "security": [
                    {
                        "apiAuth": []
                    }
                ]
            }
        },
        "/user/revoke-permissions": {
            "post": {
                "tags": [
                    "User"
                ],
                "summary": "Rovoke permissions to user",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "permission": {
                                        "type": "string"
                                    }
                                },
                                "type": "object",
                                "example": {
                                    "permission": [
                                        "edit_article",
                                        "view__article"
                                    ]
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "An example resource"
                    }
                },
                "security": [
                    {
                        "apiAuth": []
                    }
                ]
            }
        },
        "/jwt/token": {
            "post": {
                "tags": [
                    "JWT"
                ],
                "summary": "Create token",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "example": {
                                    "user_id": 1,
                                    "email": "ebislustria@gmail.com",
                                    "is_active": true
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "An example resource"
                    }
                }
            }
        },
        "/jwt/decode": {
            "post": {
                "tags": [
                    "JWT"
                ],
                "summary": "Decode token",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "example": {
                                    "token": "eyJ0eXAiOiJKV1QiLCJhbGciOiJIUzI1NiJ9.eyJpc3MiOiJFYml6Iiwic3ViIjoiTHVtZW4gVG9rZW4gSldUIiwiaWF0IjoxNTgzOTkyODQwLCJleHAiOjE1ODM5OTY0NDAsInVzZXIiOnsidXNlcl9pZCI6MiwiZW1haWwiOiJlYmlzbHVzdHJpYUBnbWFpbC5jb20iLCJpc19hY3RpdmUiOnRydWV9LCJkZXNjIjoiIn0.fHCDAVz9krhIoz8_YTgyIhunjvCFibBH_wgYjDAshQI"
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "An example resource"
                    }
                }
            }
        },
        "/cars/list": {
            "get": {
                "tags": [
                    "Cars"
                ],
                "summary": "List of Cars",
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                },
                "security": [
                    {
                        "apiAuth": []
                    }
                ]
            }
        },
        "/cars/store": {
            "post": {
                "tags": [
                    "Cars"
                ],
                "summary": "Create of Cars",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "carcompany": {
                                        "type": "string"
                                    },
                                    "name": {
                                        "type": "string"
                                    },
                                    "price": {
                                        "type": "number"
                                    }
                                },
                                "type": "object",
                                "example": {
                                    "carcompany": "Ferrari",
                                    "name": "Pista",
                                    "price": 82100000
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "An example resource"
                    }
                },
                "security": [
                    {
                        "apiAuth": []
                    }
                ]
            }
        },
        "/cars/update/{id}": {
            "put": {
                "tags": [
                    "Cars"
                ],
                "summary": "Update of Cars",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "document _id",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "carcompany": {
                                        "type": "string"
                                    },
                                    "name": {
                                        "type": "string"
                                    },
                                    "price": {
                                        "type": "number"
                                    }
                                },
                                "type": "object",
                                "example": {
                                    "carcompany": "Ferrari",
                                    "name": "Pista",
                                    "price": 82100000
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "An example resource"
                    }
                },
                "security": [
                    {
                        "apiAuth": []
                    }
                ]
            }
        },
        "/cars/destroy/{id}": {
            "delete": {
                "tags": [
                    "Cars"
                ],
                "summary": "Delete of Cars",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "document _id",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "An example resource"
                    }
                },
                "security": [
                    {
                        "apiAuth": []
                    }
                ]
            }
        },
        "/cars/detail/{id}": {
            "get": {
                "tags": [
                    "Cars"
                ],
                "summary": "Get of Cars by id",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "document _id",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "An example resource"
                    }
                },
                "security": [
                    {
                        "apiAuth": []
                    }
                ]
            }
        }
    },
    "components": {
        "securitySchemes": {
            "apiAuth": {
                "type": "http",
                "description": "Login with email and password to get the authentication token",
                "name": "Token based Based",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    }
}